+++ FilmRentalStore.API/Repositories/Interfaces/IActorRepository.cs
@@ -5,7 +5,7 @@ namespace FilmRentalStore.API.Repositories.Interfaces
     public interface IActorRepository
     {
         Task<IEnumerable<Actor>> GetAllActorsAsync();
-        Task<Actor?> GetActorByIdAsync(short id);
+        Task<Actor?> GetActorByIdAsync(int id);
         Task<bool> ActorExistsAsync(int actorId);
         Task CreateActorAsync(Actor actor);
         void Update(Actor actor);
